+++ age3-stream-agents/src/main/java/pl/edu/agh/age/compute/stream/emas/solution/Solution.java
@@ -45,7 +45,7 @@ public interface Solution<T> extends Serializable {
 	 *
 	 * @return a copy of this solution with new fitness
 	 */
-	Solution<T> updateFitness(double fitness);
+	Solution<T> withFitness(double fitness);
 
 	/**
 	 * Returns the real value of this solution.