+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.6 -- TBD
 
+Stream Agents:
+- `Solution#updateFitness` was changed to `Solution#withFitness` to follow naming used in the rest of the project (#41).
+
 Other:
 – Switch from Javaslang to Vavr after the rebranding (#40).